requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
descriptive-id==0.2.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
whitenoise==5.3.0
django-environ==0.7.0